Fix Modal bug to use react-materialize modal
