+++ client/src/components/admin/RoleManager/RoleManager.jsx
@@ -1,21 +1,31 @@
 import React, { Component } from 'react';
 import { browserHistory } from 'react-router';
 import { connect } from 'react-redux';
-import { Row, Col } from 'react-materialize';
-import ModalContainer from 'react-mf-modal/container';
-import ModalService from 'react-mf-modal';
-import { Backdrop } from 'react-mf-modal/themes/materialize';
 import PropTypes from 'prop-types';
-import { loadRoles } from '../../../actions/roleActions';
-import Modal from './Modal';
+import { Row, Col, Modal, Button } from 'react-materialize';
+import { loadRoles, saveRole } from '../../../actions/roleActions';
 import Nav from '../../Nav';
 import RoleContainer from './RoleContainer';
+
 
 /**
  * @class RoleManager
  * @extends {Component}
  */
 class RoleManager extends Component {
+  /**
+   * Creates an instance of RoleManager.
+   * @param {Object} props
+   * @memberof RoleManager
+   */
+  constructor(props) {
+    super(props);
+    this.state = {
+      role: { title: '' }
+    };
+    this.onChange = this.onChange.bind(this);
+    this.createRole = this.createRole.bind(this);
+  }
   /**
    * Hook Method
    * @returns {none} none
@@ -30,51 +40,152 @@ class RoleManager extends Component {
     }
   }
 
-   handleModal = () => {
-     ModalService.open(<Modal />);
-   }
+  /**
+   * Function to handle onchange event on text input
+   * @param {Object} e: browser onchange event
+   * @memberOf RoleManager
+   * @returns {none} Updates state
+   */
+  onChange(e) {
+    const title = e.target.name;
+    const role = this.state.role;
+    role[title] = e.target.value;
+    this.setState({ role });
+  }
+
+  /**
+   * Function to handle creation of new roles
+   * @param {Object} e: browser event
+   * @memberOf RoleManager
+   * @returns {none} handles form onChange event
+   */
+  createRole(e) {
+    e.preventDefault();
+    const currentRoles = this.props.roles.map(role => role.title);
+    const newRole = this.state.role.title;
+    if (currentRoles.includes(newRole)) {
+      swal({  //eslint-disable-line
+        title: 'Invalid Operation',
+        text: 'This Role already exist',
+        type: 'warning',
+        showCancelButton: false,
+        confirmButtonColor: '#DD6B55',
+        confirmButtonText: 'Undo',
+        closeOnConfirm: true
+      });
+    } else if (newRole === undefined) {
+        swal({  //eslint-disable-line
+        title: 'Invalid Operation',
+        text: 'Please Enter a role title',
+        type: 'warning',
+        showCancelButton: false,
+        confirmButtonColor: '#DD6B55',
+        confirmButtonText: 'Undo',
+        closeOnConfirm: true
+      });
+    } else {
+      this.props.saveRole(this.state.role)
+        .then(() => {
+          swal( //eslint-disable-line
+            'Nice',
+            'Role saved succefully',
+            'success'
+          );
+        })
+        .catch(() => {
+          swal({  //eslint-disable-line
+            title: 'Invalid Operation',
+            text: 'Please Enter a valid role title',
+            type: 'warning',
+            showCancelButton: false,
+            confirmButtonColor: '#DD6B55',
+            confirmButtonText: 'Undo',
+            closeOnConfirm: true
+          });
+        });
+    }
+  }
   /**
    * @returns {Object} Jsx
    * @memberOf RoleManager
    */
-   render() {
-     return (
-       <div>
-         <Nav />
-         <div>
-           <Row>
-             <Col m={1} />
-             <Col m={10}>
+  render() {
+    return (
+      <div>
+        <Nav />
+        <div>
+          <Row>
+            <Col m={1} />
+            <Col m={10}>
 
-               <RoleContainer data={this.props.roles} />
-               <ModalContainer backdropComponent={Backdrop}>
-                 <a // eslint-disable-line
-                   onClick={this.handleModal}
-                   className="btn-floating btn-large waves-effect waves-light red"
-                 >
-                   <i
-                     className="material-icons"
-                     data-tool-tip="add new role"
-                   >add</i>
-                 </a>
-               </ModalContainer>
-             </Col>
-             <Col m={1} />
-           </Row>
-         </div>
-       </div>
-     );
-   }
+              <RoleContainer data={this.props.roles} />
+              <Modal
+                actions={
+                  [
+                    <Button
+                      style={{ marginLeft: `${2}em` }}
+                      className="btn-save"
+                      waves="light"
+                      modal="close"
+                      flat
+                    >Close
+                    </Button>,
+                    <Button
+                      id="save-role"
+                      waves="light"
+                      flat
+                      className="btn-save"
+                      onClick={this.createRole}
+                    >Save
+                    </Button>
+                  ]
+                }
+                trigger={
+                  <a
+                    className="btn-floating btn-large waves-effect waves-light red"
+                  >
+                    <i
+                      className="material-icons"
+                      data-tool-tip="add new role"
+                    >add</i>
+                  </a>
+                }
+              >
+                <Row>
+                  <div className="col s8 push-s2">
+                    <div>
+                      <label htmlFor="newRole">New Role Title</label>
+                      <input
+                        id="role"
+                        type="text"
+                        icon="person_otline"
+                        placeholder="Add a role title here"
+                        onChange={this.onChange}
+                        name="title"
+                        value={this.state.role.title}
+                      />
+                    </div>
+                  </div>
+                </Row>
+              </Modal>
+            </Col>
+            <Col m={1} />
+          </Row>
+        </div>
+      </div>
+    );
+  }
 }
 RoleManager.propTypes = {
   user: PropTypes.object.isRequired,
   loadRoles: PropTypes.func.isRequired,
-  roles: PropTypes.array.isRequired
+  roles: PropTypes.array.isRequired,
+  saveRole: PropTypes.func.isRequired
 };
 const mapStateToProps = state => ({
-  user: state.auth.user,
-  roles: state.manageRoles.roles
+  roles: state.manageRoles.roles,
+  user: state.auth.user
 });
 
 export default connect(mapStateToProps,
-  { loadRoles })(RoleManager);
+  { loadRoles, saveRole })(RoleManager);
