+++ client/src/index.jsx
@@ -6,9 +6,6 @@ import axios from 'axios';
 import jwtDecode from 'jwt-decode';
 
 import 'material-ui';
-import 'materialize-css'; //eslint-disable-line
-import 'materialize-css/dist/css/materialize.css';
-import 'materialize-css/dist/js/materialize'; //eslint-disable-line
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 
 import { Provider } from 'react-redux';
